Some code clean up.
