+++ Big Neon/Screens/GuestListViewController/EventViewModel.swift
@@ -23,7 +23,6 @@ final class EventViewModel {
     var guestCoreData: [GuestData] = []
     var guestCoreDataSearchResults: [GuestData] = []
     var guestSearchResults: [RedeemableTicket] = []
-    var checkinService: CheckinService = CheckinService()
     
     let dataStack = DataStack(modelName: "Big Neon")
 
@@ -53,8 +52,6 @@ final class EventViewModel {
      Delete Core Data Guests
      */
     func deleteAllData(_ entity: String) {
-
-//        let appDel =  UIApplication.shared.delegate as! AppDelegate
         let context:NSManagedObjectContext = CoreDataStack.sharedInstance.persistentContainer.viewContext
 
         let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: entity)
@@ -107,54 +104,6 @@ final class EventViewModel {
                 }
             }
         }
-    }
-    
-    /*
-     Save Scanned Ticket In Core Data
-    */
-    func saveScannedTicketInCoreDataWith(array: [[String: AnyObject]]) {
-        _ = array.map{self.createScannedTicketFrom(dictionary: $0)}
-        do {
-            try CoreDataStack.sharedInstance.persistentContainer.viewContext.save()
-        } catch let error {
-            print(error)
-        }
-    }
-    
-    func createScannedTicketFrom(dictionary: [String: AnyObject]) -> NSManagedObject? {
-        
-        let context = CoreDataStack.sharedInstance.persistentContainer.viewContext
-        if let scannedTicketEntity = NSEntityDescription.insertNewObject(forEntityName: SCANNED_TICKET_ENTITY_NAME, into: context) as? ScannedTicketData {
-            scannedTicketEntity.id = dictionary["id"] as? String
-            scannedTicketEntity.event_id = dictionary["event_id"] as? String
-            return scannedTicketEntity
-        }
-        return nil
-    }
-    
-    
-    /*
-     Fetch guests from data and sync up to network
-     */
-    func fetchScannedLocalGuests(completion: @escaping(Bool) -> Void) {
-        let context:NSManagedObjectContext = CoreDataStack.sharedInstance.persistentContainer.viewContext
-        let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: SCANNED_TICKET_ENTITY_NAME)
-        
-        do {
-            if let scannedTicket = try context.fetch(fetchRequest) as? [ScannedTicketData] {
-                for ticket in scannedTicket {
-                    self.checkinService.automaticallyCheckin(ticketID: ticket.id!, eventID: ticket.event_id!) { (scanFeedBack, errorString, redeemedTicket) in
-                        print("Scan Feedback: \(scanFeedBack), \(errorString), \(redeemedTicket)")
-                    }
-                    
-                    //  Delete the Ticket Value Here from Core Data
-                    
-                }
-            }
-        } catch let error {
-            print("Error fetching all data in \(SCANNED_GUEST_ENTITY_NAME):", error)
-        }
-        
     }
     
     /*
